backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlparse>=0.5.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
pandas>=2.2.0
lxml>=5.2.0
PyYAML>=6.0.1
google-generativeai>=0.7.0
python-dotenv>=1.0.1
